+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.4.0
+- Defer `resolveBEMModifiers` to `truthyStringsKeys`.
+- **Breaking change:** Remove exported `resolveBEMModifiers` function.
+
 ## v0.3.0
 - Install [`truthy-keys`](https://www.npmjs.com/package/truthy-keys) dependency.
 - **Breaking change:** Remove exported `keys` function.